moved Rabbit config, added consumer functionality
